move card search to a module and delete old card select
