+++ client/visual/shortcut.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+const Shortcut = styled.span`
+	font-size: 0.7em;
+	opacity: 0.5;
+	border-width: 1px;
+	border-style: solid;
+	border-radius: 2px;
+	padding: 0.25em;
+	margin-right: 0.5em;
+`
+
+export default Shortcut
